+++ src/screens/auth/BackupSeedPhrase/styles.js
