+++ src/screens/auth/CreatePassword/index.js
