+++ src/screens/auth/ImportSeedPhrase/index.js
