+++ src/screens/error/ConnectionError/index.js
@@ -1,9 +1,10 @@
 import React from 'react';
 import { View, Image } from 'react-native';
 
-import { ERROR_TYPES } from '../../components/common/ErrorState/constants';
-import ErrorState from '../../components/common/ErrorState';
-import PlugLogo from '../../assets/icons/il_white_plug.png';
+import { ERROR_TYPES } from '@commonComponents/ErrorState/constants';
+import PlugLogo from '@assets/icons/il_white_plug.png';
+import ErrorState from '@commonComponents/ErrorState';
+
 import styles from './styles';
 
 function ConnectionError() {
