+++ src/screens/error/ConnectionError/styles.js
@@ -1,6 +1,6 @@
 import { StyleSheet } from 'react-native';
 
-import { Colors } from '../../constants/theme';
+import { Colors } from '@constants/theme';
 
 export default StyleSheet.create({
   container: {
