+++ src/screens/flows/Deposit/components/DepositDivider/index.js
