+++ src/screens/flows/Deposit/components/DepositDivider/styles.js
