+++ src/screens/flows/Deposit/components/IDDetails/styles.js
@@ -1,6 +1,6 @@
 import { StyleSheet } from 'react-native';
 
-import { FontStyles } from '../../../../constants/theme';
+import { FontStyles } from '@constants/theme';
 
 export default StyleSheet.create({
   container: {
