+++ src/screens/flows/Send/components/TokenSection/index.js
