+++ src/screens/tabs/NFTs/screens/NftDetail/components/Section/index.js
@@ -1,0 +1,15 @@
+import React from 'react';
+import { View, Text } from 'react-native';
+
+import { FontStyles } from '@constants/theme';
+
+import styles from './styles';
+
+const Section = ({ title, children, style }) => (
+  <View style={[styles.section, style]}>
+    <Text style={FontStyles.Subtitle}>{title}</Text>
+    <View style={styles.content}>{children}</View>
+  </View>
+);
+
+export default Section;
