+++ src/screens/tabs/NFTs/screens/NftDetail/index.js
@@ -20,11 +20,11 @@ import { FontStyles } from '@constants/theme';
 import Badge from '@commonComponents/Badge';
 import { isIos } from '@constants/platform';
 import useGetType from '@hooks/useGetType';
+import Send from '@screens/flows/Send';
 import Modal from '@components/modal';
 
 import Section from './components/Section';
 import styles from './styles';
-import Send from '../Send';
 
 const NftDetail = ({ modalRef, handleClose, selectedNFT, ...props }) => {
   const isCapCrowns = selectedNFT?.collection === 'CAP Crowns';
