+++ src/screens/tabs/NFTs/screens/NftDetail/styles.js
