+++ src/screens/tabs/Profile/screens/Accounts/index.js
