+++ src/screens/tabs/Profile/screens/RevealSeedPhrase/index.js
