+++ src/screens/tabs/Profile/screens/RevealSeedPhrase/styles.js
@@ -1,6 +1,6 @@
 import { StyleSheet } from 'react-native';
 
-import { FontStyles } from '../../constants/theme';
+import { FontStyles } from '@constants/theme';
 
 export default StyleSheet.create({
   container: {
