+++ src/screens/tabs/Profile/screens/Settings/components/SettingItem/styles.js
@@ -1,0 +1,21 @@
+import { StyleSheet } from 'react-native';
+
+export default StyleSheet.create({
+  container: {
+    paddingHorizontal: 20,
+    paddingVertical: 20,
+  },
+  icon: {
+    fontSize: 18,
+  },
+  chevron: {
+    marginLeft: 'auto',
+    alignSelf: 'center',
+  },
+  iconName: {
+    marginRight: 5,
+  },
+  name: {
+    marginBottom: 5,
+  },
+});
